requirements.txt
colorama==0.4.6
loguru==0.7.2
lxml==5.1.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-docx==1.1.0
typing_extensions==4.9.0
win32-setctime==1.1.0
requests==2.31.0
openpyxl==3.1.2
beautifulsoup4==4.12.3